Integrated and replaced old camera
